third_party/blink/web_tests/fast/dom/HTMLElement/iscontenteditable-designmodeon-expected.txt

isContentEditable should not change if its contentEditable is set, no matter whether document.designMode is enabled or not.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

document.designMode = "on"
PASS document.getElementById("div").getAttribute("contentEditable") is "false"
PASS document.getElementById("div").contentEditable is "false"
PASS document.getElementById("div").isContentEditable is true
PASS window.getComputedStyle(div, "").getPropertyValue("-webkit-user-modify") is "read-only"
PASS document.getElementById("div2").getAttribute("contentEditable") is "false"
PASS document.getElementById("div2").contentEditable is "false"
PASS document.getElementById("div2").isContentEditable is true
PASS window.getComputedStyle(div2, "").getPropertyValue("-webkit-user-modify") is "read-only"
PASS document.getElementById("p").getAttribute("contentEditable") is "true"
PASS document.getElementById("p").contentEditable is "true"
PASS document.getElementById("p").isContentEditable is true
PASS window.getComputedStyle(p, "").getPropertyValue("-webkit-user-modify") is "read-write"
PASS document.getElementById("p2").getAttribute("contentEditable") is "true"
PASS document.getElementById("p2").contentEditable is "true"
PASS document.getElementById("p2").isContentEditable is true
PASS window.getComputedStyle(p2, "").getPropertyValue("-webkit-user-modify") is "read-write"
PASS successfullyParsed is true

TEST COMPLETE
